refactor: deleteAll 메서드 제거
