+++ src/main/java/wooteco/subway/service/LineService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.dao.JdbcLineDao;
+import wooteco.subway.dao.LineDao;
 import wooteco.subway.domain.Line;
 import wooteco.subway.dto.LineRequest;
 import wooteco.subway.dto.LineResponse;
@@ -16,17 +16,17 @@ import wooteco.subway.exception.NotFoundException;
 @Transactional(readOnly = true)
 public class LineService {
 
-    private final JdbcLineDao jdbcLineDao;
+    private final LineDao lineDao;
 
-    public LineService(JdbcLineDao jdbcLineDao) {
-        this.jdbcLineDao = jdbcLineDao;
+    public LineService(LineDao lineDao) {
+        this.lineDao = lineDao;
     }
 
     @Transactional
     public LineResponse create(LineRequest lineRequest) {
         Line line = new Line(lineRequest.getName(), lineRequest.getColor());
         validateDuplicateNameAndColor(line.getName(), line.getColor());
-        return LineResponse.from(jdbcLineDao.save(line));
+        return LineResponse.from(lineDao.save(line));
     }
 
     public LineResponse showById(Long id) {
@@ -34,7 +34,7 @@ public class LineService {
     }
 
     public List<LineResponse> showAll() {
-        return jdbcLineDao.findAll().stream()
+        return lineDao.findAll().stream()
             .map(LineResponse::from)
             .collect(Collectors.toList());
     }
@@ -44,21 +44,21 @@ public class LineService {
         validateDuplicateNameAndColor(request.getName(), request.getColor());
         Line line = findBy(id);
         line.update(request.getName(), request.getColor());
-        jdbcLineDao.modifyById(id, line);
+        lineDao.modifyById(id, line);
     }
 
     @Transactional
     public void removeById(Long id) {
-        jdbcLineDao.deleteById(id);
+        lineDao.deleteById(id);
     }
 
     private Line findBy(Long id) {
-        return jdbcLineDao.findById(id)
+        return lineDao.findById(id)
             .orElseThrow(() -> new NotFoundException("조회하려는 id가 존재하지 않습니다."));
     }
 
     private void validateDuplicateNameAndColor(String name, String color) {
-        if (jdbcLineDao.existByNameAndColor(name, color)) {
+        if (lineDao.existByNameAndColor(name, color)) {
             throw new BadRequestException("노선이 이름과 색상은 중복될 수 없습니다.");
         }
     }
