+++ src/main/java/wooteco/subway/service/StationService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import wooteco.subway.dao.JdbcStationDao;
+import wooteco.subway.dao.StationDao;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
@@ -14,22 +14,22 @@ import wooteco.subway.exception.BadRequestException;
 @Transactional(readOnly = true)
 public class StationService {
 
-    private final JdbcStationDao jdbcStationDao;
+    private final StationDao stationDao;
 
-    public StationService(JdbcStationDao jdbcStationDao) {
-        this.jdbcStationDao = jdbcStationDao;
+    public StationService(StationDao stationDao) {
+        this.stationDao = stationDao;
     }
 
     @Transactional
     public StationResponse create(StationRequest stationRequest) {
         Station station = new Station(stationRequest.getName());
         validateDuplicateName(station);
-        Station savedStation = jdbcStationDao.save(station);
+        Station savedStation = stationDao.save(station);
         return new StationResponse(savedStation.getId(), savedStation.getName());
     }
 
     public List<StationResponse> showAll() {
-        List<Station> stations = jdbcStationDao.findAll();
+        List<Station> stations = stationDao.findAll();
         return stations.stream()
             .map(value -> new StationResponse(value.getId(), value.getName()))
             .collect(Collectors.toList());
@@ -37,11 +37,11 @@ public class StationService {
 
     @Transactional
     public void removeById(Long id) {
-        jdbcStationDao.deleteById(id);
+        stationDao.deleteById(id);
     }
 
     private void validateDuplicateName(Station station) {
-        if (jdbcStationDao.existByName(station.getName())) {
+        if (stationDao.existByName(station.getName())) {
             throw new BadRequestException("지하철 역 이름은 중복될 수 없습니다.");
         }
     }
