+++ src/test/java/wooteco/subway/dao/LineDaoTest.java
@@ -1,96 +1,74 @@
 package wooteco.subway.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
+import org.springframework.context.annotation.Import;
 import wooteco.subway.domain.Line;
 
-class LineDaoTest {
+@JdbcTest
+@Import(LineDao.class)
+public class LineDaoTest {
 
-    @BeforeEach
-    void setUp() {
-        LineDao.deleteAll();
-    }
+    @Autowired
+    private LineDao lineDao;
 
     @Test
-    @DisplayName("노선을 저장한다.")
+    @DisplayName("노선 저장")
     void save() {
-        Line line = LineDao.save(new Line("2호선", "bg-red-600"));
-        assertThat(line.getName()).isEqualTo("2호선");
+        Line line = new Line("1호선", "blue");
+        Line savedLine = lineDao.save(line);
+        assertThat(savedLine.getId()).isNotNull();
+        assertThat(savedLine.getName()).isEqualTo("1호선");
     }
 
     @Test
-    @DisplayName("노선을 id로 조회한다.")
+    @DisplayName("지하철 역 이름 중복 여부 조회")
+    void duplicateName() {
+        Line line = new Line("1호선", "blue");
+        lineDao.save(line);
+        assertThat(lineDao.existByNameAndColor("1호선", "blue")).isTrue();
+    }
+
+    @Test
+    @DisplayName("id로 노선 조회")
     void findById() {
-        Line line = LineDao.save(new Line("2호선", "bg-red-600"));
-        Line findLine = LineDao.findById(line.getId());
-        assertThat(findLine.getName()).isEqualTo("2호선");
-    }
-
-    @ParameterizedTest(name = "name : {0}, color : {1}")
-    @CsvSource({"1호선, black", "1호선, blue", "2호선, black"})
-    void duplicateNameAndColor(String name, String color) {
-        LineDao.save(new Line(name, color));
-        assertThatIllegalArgumentException()
-            .isThrownBy(() -> LineDao.save(new Line("1호선", "black")))
-            .withMessage("노선의 이름은 중복될 수 없습니다.");
+        Line line = lineDao.save(new Line("1호선", "blue"));
+        Line findLine = lineDao.findById(line.getId()).get();
+        assertThat(findLine.getId()).isNotNull();
+        assertThat(findLine.getName()).isEqualTo("1호선");
     }
 
     @Test
-    @DisplayName("노선을 저장할 때 id가 1씩 증가한다.")
-    void increaseId() {
-        Line line1 = LineDao.save(new Line("2호선", "bg-red-600"));
-        Line line2 = LineDao.save(new Line("3호선", "bg-blue-600"));
-        assertThat(line2.getId() - line1.getId()).isEqualTo(1L);
+    @DisplayName("노선 전체 조회")
+    void findAll() {
+        Line line1 = new Line("1호선", "blue");
+        Line line2 = new Line("2호선", "red");
+        lineDao.save(line1);
+        lineDao.save(line2);
+        List<Line> liens = lineDao.findAll();
+        assertThat(liens).hasSize(2);
     }
 
     @Test
-    @DisplayName("노선 목록을 조회한다.")
-    void showLineList() {
-        LineDao.save(new Line("1호선", "blue"));
-        LineDao.save(new Line("2호선", "green"));
-        List<Line> values = LineDao.findAll();
-        assertThat(values).hasSize(2);
-    }
-
-    @Test
-    @DisplayName("없는 노선 목록을 조회할 경우 예외를 발생시킨다.")
-    void notFindById() {
-        assertThatIllegalArgumentException()
-            .isThrownBy(() -> LineDao.findById(1L))
-            .withMessage("조회하려는 id가 없습니다.");
-    }
-
-    @Test
-    @DisplayName("id로 노선을 수정한다.")
+    @DisplayName("id로 노선 수정")
     void modifyById() {
-        Line savedLine = LineDao.save(new Line("1호선", "blue"));
-        LineDao.modify(savedLine.getId(), "3호선", "green");
-
-        Line modifiedLine = LineDao.findById(savedLine.getId());
-        assertThat(modifiedLine.getName()).isEqualTo("3호선");
-    }
-
-    @ParameterizedTest(name = "name : {0}, color : {1}")
-    @CsvSource({"1호선, black", "1호선, red"})
-    void modifyDuplicateNameAndColor(String name, String color) {
-        Line line = LineDao.save(new Line("1호선", "black"));
-        assertThatIllegalArgumentException()
-            .isThrownBy(() -> LineDao.modify(line.getId(), name, color))
-            .withMessage("노선의 이름과 색깔은 중복될 수 없습니다.");
+        Line savedLine = lineDao.save(new Line("1호선", "blue"));
+        lineDao.modifyById(savedLine.getId(), new Line("2호선", "red"));
+        Line updateLine = lineDao.findById(savedLine.getId()).get();
+        assertThat(updateLine.getName()).isEqualTo("2호선");
+        assertThat(updateLine.getColor()).isEqualTo("red");
     }
 
     @Test
-    @DisplayName("id로 노선을 삭제한다.")
+    @DisplayName("id로 노선 삭제")
     void deleteById() {
-        Line savedLine = LineDao.save(new Line("1호선", "blue"));
-        LineDao.deleteById(savedLine.getId());
-        assertThat(LineDao.findAll()).hasSize(0);
+        Line savedLine = lineDao.save(new Line("1호선", "blue"));
+        lineDao.deleteById(savedLine.getId());
+        assertThat(lineDao.findAll()).hasSize(0);
     }
 
 }
